% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_allocatoR_refactored.R
\name{check_significance}
\alias{check_significance}
\title{Check for Significant Covariates in Output}
\usage{
check_significance(output)
}
\arguments{
\item{output}{A list or similar object expected to contain a \code{results} data frame. The \code{results} data frame should have at least two columns: \code{covariate} for the name of the covariate and \code{p_value} for the associated p-value of some statistical test.}
}
\value{
A data frame listing the covariates that have been found to be statistically significant after adjusting their p-values for multiple testing. If no covariates are found to be significant, the function will return an empty data frame. The return data frame includes columns for the covariate name and the count of significant instances, although the count will typically be one for each listed covariate unless the input \code{output} structure allows for duplicate covariate names.
}
\description{
This function evaluates the significance of covariates based on adjusted p-values. It applies the Bonferroni correction to the p-values associated with each covariate and identifies those that are statistically significant.
We \emph{do not} recommend using this function to conduct hypothesis testing, but instead only as a sanity check to identify which covariates may potentially be problematic.
}
\details{
The function first retrieves the \code{results} data frame from the provided \code{output} object. It then calculates adjusted p-values using the Bonferroni correction method. Covariates with adjusted p-values less than 0.05 are considered statistically significant. The function outputs the names of significant covariates and their counts.
}
\examples{
my_data = simulate_data(n_samples = 100)

allocated_data <- allocate_samples(data = my_data,
                                   id_column = "sample_id",
                                   method = "simulated_annealing",
                                   covariates = c("age_at_baseline", "bmi_at_baseline", "sex"),
                                   batch_size = 13)

significant_covariates <- check_significance(allocated_data)

}
